+++ CHANGELOG.md
@@ -17,6 +17,7 @@ maddy uses [semver versioning](https://semver.org/).
 * ![**CHANGED**](https://img.shields.io/badge/-CHANGED-%23e90) Updated google test to v1.16.0.
 * ![**ADDED**](https://img.shields.io/badge/-ADDED-%23099) clang-format
 * ![**ADDED**](https://img.shields.io/badge/-ADDED-%23099) automatic update dependencies ci
+* ![**FIXED**](https://img.shields.io/badge/-FIXED-%23090) `(This is a [link](/ABC/some file) (the URL will include this).)` should not put the later parenthesis into the link url.
 
 ## version 1.3.0 2023-08-26